perf lock contention: Fix a build error on 32-bit
Bugzilla: https://bugzilla.redhat.com/2123229

upstream
========
commit 433b31fa00797a2a6205a023e9345f2c5e7896b6
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Sep 26 14:56:38 2022 -0700

description
===========
It was reported that it failed to build the BPF lock contention skeleton
on 32 bit arch due to the size of long.  The lost count is used only for
reporting errors due to lack of stackmap space through bad_hist which
type is 'int'.  Let's use int type then.

Fixes: 6d499a6b3d90277d ("perf lock: Print the number of lost entries for BPF")
    Reported-by: Jiri Slaby <jirislaby@kernel.org>
    Cc: Adrian Hunter <adrian.hunter@intel.com>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Song Liu <songliubraving@fb.com>
    Link: http://lore.kernel.org/lkml/20220926215638.3931222-1-namhyung@kernel.org

Conflicts:
==========
Not-yet-backported 96532a83ee8e ("perf lock contention: Allow to change
stack depth and skip") caused different context.

+++ tools/perf/util/bpf_skel/lock_contention.bpf.c
@@ -74,7 +74,7 @@ int has_cpu;
 int has_task;
 
 /* error stat */
-unsigned long lost;
+int lost;
 
 static inline int can_record(void)
 {
